+++ src/main/java/com/altama/forecast/interfaces/web/facade/dto/m_productDTO/M_productDTOBuilder.java
@@ -95,7 +95,7 @@ public class M_productDTOBuilder {
     private BigDecimal palletizevalue;
     private String movingtype;
     private String movingtypesales;
-    private short interface_flag;
+    private int interface_flag;
     private BigDecimal weight_ctn;
     private BigDecimal qtylimitpo;
     private BigDecimal length_ctn;
@@ -537,7 +537,7 @@ public class M_productDTOBuilder {
         return this;
     }
 
-    public M_productDTOBuilder setInterface_flag(short interface_flag) {
+    public M_productDTOBuilder setInterface_flag(int interface_flag) {
         this.interface_flag = interface_flag;
         return this;
     }
